Add functionality for Delete task
