+++ src/main/java/manageezpz/logic/commands/DeleteEmployeeCommand.java
@@ -1,0 +1,53 @@
+package manageezpz.logic.commands;
+
+import static java.util.Objects.requireNonNull;
+
+import java.util.List;
+
+import manageezpz.commons.core.Messages;
+import manageezpz.commons.core.index.Index;
+import manageezpz.logic.commands.exceptions.CommandException;
+import manageezpz.model.Model;
+import manageezpz.model.person.Person;
+
+/**
+ * Deletes a person identified using it's displayed index from the address book.
+ */
+public class DeleteEmployeeCommand extends Command {
+
+    public static final String COMMAND_WORD = "delete";
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD
+            + ": Deletes the person identified by the index number used in the displayed person list.\n"
+            + "Parameters: INDEX (must be a positive integer)\n"
+            + "Example: " + COMMAND_WORD + " 1";
+
+    public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
+
+    private final Index targetIndex;
+
+    public DeleteEmployeeCommand(Index targetIndex) {
+        this.targetIndex = targetIndex;
+    }
+
+    @Override
+    public CommandResult execute(Model model) throws CommandException {
+        requireNonNull(model);
+        List<Person> lastShownList = model.getFilteredPersonList();
+
+        if (targetIndex.getZeroBased() >= lastShownList.size()) {
+            throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
+        }
+
+        Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+        model.deletePerson(personToDelete);
+        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, personToDelete));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof DeleteEmployeeCommand // instanceof handles nulls
+                && targetIndex.equals(((DeleteEmployeeCommand) other).targetIndex)); // state check
+    }
+}
