+++ src/main/java/manageezpz/logic/parser/AddressBookParser.java
@@ -12,7 +12,7 @@ import manageezpz.logic.commands.AddEventTaskCommand;
 import manageezpz.logic.commands.AddTodoTaskCommand;
 import manageezpz.logic.commands.ClearCommand;
 import manageezpz.logic.commands.Command;
-import manageezpz.logic.commands.DeleteCommand;
+import manageezpz.logic.commands.DeleteTaskCommand;
 import manageezpz.logic.commands.EditCommand;
 import manageezpz.logic.commands.ExitCommand;
 import manageezpz.logic.commands.FindCommand;
@@ -53,8 +53,8 @@ public class AddressBookParser {
         case EditCommand.COMMAND_WORD:
             return new EditCommandParser().parse(arguments);
 
-        case DeleteCommand.COMMAND_WORD:
-            return new DeleteCommandParser().parse(arguments);
+        case DeleteTaskCommand.COMMAND_WORD:
+            return new DeleteTaskCommandParser().parse(arguments);
 
         case ClearCommand.COMMAND_WORD:
             return new ClearCommand();
