--- src/main/java/manageezpz/logic/parser/DeleteCommandParser.java
@@ -1,29 +1,0 @@
-package manageezpz.logic.parser;
-
-import static manageezpz.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-
-import manageezpz.commons.core.index.Index;
-import manageezpz.logic.commands.DeleteCommand;
-import manageezpz.logic.parser.exceptions.ParseException;
-
-/**
- * Parses input arguments and creates a new DeleteCommand object
- */
-public class DeleteCommandParser implements Parser<DeleteCommand> {
-
-    /**
-     * Parses the given {@code String} of arguments in the context of the DeleteCommand
-     * and returns a DeleteCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public DeleteCommand parse(String args) throws ParseException {
-        try {
-            Index index = ParserUtil.parseIndex(args);
-            return new DeleteCommand(index);
-        } catch (ParseException pe) {
-            throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE), pe);
-        }
-    }
-
-}
