+++ src/main/java/manageezpz/logic/parser/DeleteEmployeeCommandParser.java
@@ -1,0 +1,29 @@
+package manageezpz.logic.parser;
+
+import static manageezpz.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import manageezpz.commons.core.index.Index;
+import manageezpz.logic.commands.DeleteEmployeeCommand;
+import manageezpz.logic.parser.exceptions.ParseException;
+
+/**
+ * Parses input arguments and creates a new DeleteCommand object
+ */
+public class DeleteEmployeeCommandParser implements Parser<DeleteEmployeeCommand> {
+
+    /**
+     * Parses the given {@code String} of arguments in the context of the DeleteCommand
+     * and returns a DeleteCommand object for execution.
+     * @throws ParseException if the user input does not conform the expected format
+     */
+    public DeleteEmployeeCommand parse(String args) throws ParseException {
+        try {
+            Index index = ParserUtil.parseIndex(args);
+            return new DeleteEmployeeCommand(index);
+        } catch (ParseException pe) {
+            throw new ParseException(
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteEmployeeCommand.MESSAGE_USAGE), pe);
+        }
+    }
+
+}
