+++ src/test/java/manageezpz/logic/parser/DeleteEmployeeCommandParserTest.java
@@ -1,0 +1,32 @@
+package manageezpz.logic.parser;
+
+import static manageezpz.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static manageezpz.logic.parser.CommandParserTestUtil.assertParseFailure;
+import static manageezpz.logic.parser.CommandParserTestUtil.assertParseSuccess;
+import static manageezpz.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
+
+import org.junit.jupiter.api.Test;
+
+import manageezpz.logic.commands.DeleteEmployeeCommand;
+
+/**
+ * As we are only doing white-box testing, our test cases do not cover path variations
+ * outside of the DeleteCommand code. For example, inputs "1" and "1 abc" take the
+ * same path through the DeleteCommand, and therefore we test only one of them.
+ * The path variation for those two cases occur inside the ParserUtil, and
+ * therefore should be covered by the ParserUtilTest.
+ */
+public class DeleteEmployeeCommandParserTest {
+
+    private DeleteEmployeeCommandParser parser = new DeleteEmployeeCommandParser();
+
+    @Test
+    public void parse_validArgs_returnsDeleteCommand() {
+        assertParseSuccess(parser, "1", new DeleteEmployeeCommand(INDEX_FIRST_PERSON));
+    }
+
+    @Test
+    public void parse_invalidArgs_throwsParseException() {
+        assertParseFailure(parser, "a", String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteEmployeeCommand.MESSAGE_USAGE));
+    }
+}
